drop structures

+++ database/migrations/2018_12_03_193037_drop_fleet_end_column_fleets_table.php
@@ -13,9 +13,11 @@ class DropFleetEndColumnFleetsTable extends Migration
      */
     public function up()
     {
-        Schema::table('Fleets', function($table) {
-            $table->dropColumn('fleet_end');
-        });
+        if(Schema::hasColumn('Fleets', 'fleet_end')) {
+            Schema::table('Fleets', function($table) {
+                $table->dropColumn('fleet_end');
+            });
+        }
     }
 
     /**
